+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 * Changed namespace from `ContentfulBootstrap` to `Contentful::Bootstrap` to mimic other libraries
 * Changed repository name from `contentful_bootstrap.rb` to `contentful-bootstrap.rb` to mimic other libraries
 * Delivery API Token will always get created when using `contentful_bootstrap` commands to create a space
+* Configuration now read from `~/.contentfulrc`
+* Tool now requests user to allow to write configuration files
 
 ### Added
 * `contentful-bootstrap.rb` version number to API Token description
+* `inifile` as runtime dependency
+* Add optional `--config CONFIG_PATH` parameter to commands
 
 ## v0.0.7
 ### Fixed